* REFACTOR: Create ProfileService * ADD: Unit test send email
