+++ backend/src/test/java/com/huellapositiva/unit/UpdateVolunteerProfileActionShould.java
@@ -3,6 +3,7 @@ package com.huellapositiva.unit;
 import com.huellapositiva.application.dto.ProfileDto;
 import com.huellapositiva.domain.actions.UpdateVolunteerProfileAction;
 import com.huellapositiva.domain.service.EmailCommunicationService;
+import com.huellapositiva.domain.service.ProfileService;
 import com.huellapositiva.infrastructure.orm.repository.JpaVolunteerRepository;
 import com.huellapositiva.util.ProfileDtoDataEntry;
 import com.huellapositiva.util.TestData;
@@ -24,31 +25,23 @@ import static org.mockito.Mockito.verify;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ExtendWith(MockitoExtension.class)
 @Import(TestData.class)
-public class UpdateVolunteerProfileActionShould {
+class UpdateVolunteerProfileActionShould {
 
-    @Autowired
     private UpdateVolunteerProfileAction updateVolunteerProfileAction;
-
-    @Autowired
-    private TestData testData;
 
     @Mock
     private EmailCommunicationService emailCommunicationService;
 
-    @Autowired
-    private JpaVolunteerRepository jpaVolunteerRepository;
+    @Mock
+    private ProfileService profileService;
 
     @BeforeEach
     void beforeEach(){
-        //updateVolunteerProfileAction = new UpdateVolunteerProfileAction(emailCommunicationService);
+        updateVolunteerProfileAction = new UpdateVolunteerProfileAction(profileService, emailCommunicationService);
     }
 
     @Test
     void send_change_email(){
-        /*updateVolunteerProfileAction = new UpdateVolunteerProfileAction(emailCommunicationService);
-        testData.createVolunteerWithProfile(DEFAULT_EMAIL, DEFAULT_PASSWORD);
-        System.out.println("hola");
-
         ProfileDto profileDto = ProfileDto.builder()
                 .name("nombre")
                 .surname("apellido")
@@ -59,8 +52,6 @@ public class UpdateVolunteerProfileActionShould {
                 .island("Fuerteventura")
                 .build();
         updateVolunteerProfileAction.execute(profileDto, DEFAULT_EMAIL);
-        verify(emailCommunicationService).sendMessageEmailChanged(any());*/
-
+        verify(emailCommunicationService).sendMessageEmailChanged(any());
     }
-
 }
